+++ custom_components/growatt_modbus/const.py
@@ -9,6 +9,7 @@ MANUFACTURER = "Growatt"
 
 PLATFORMS: list[Platform] = [
     Platform.BINARY_SENSOR,
+    Platform.BUTTON,
     Platform.NUMBER,
     Platform.SELECT,
     Platform.SENSOR,